alterações realizadas
