+++ src/app/pages/access-and-security/access-and-security.page.html
@@ -17,7 +17,7 @@
       <ion-title class="font-tommy-bold text-base text-[#393939] p-0">
         Usuários
       </ion-title>
-      <ion-button
+      <ion-button (click)="openUserCreateModal()"
         class="add-user font-tommy-bold text-xs normal-case min-h-[30px] m-0 box-shadow"
       >
         <svg
