+++ src/app/pages/access-and-security/access-and-security.page.ts
@@ -1,7 +1,34 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { IonContent, IonTitle, IonText, IonList, IonListHeader, IonItem, IonButton, IonHeader, IonToolbar } from '@ionic/angular/standalone';
+import { IonContent, IonTitle, IonText, IonList, IonListHeader, IonItem, IonButton, IonHeader, IonToolbar, ModalController } from '@ionic/angular/standalone';
+import { UserCreateComponent } from '../../components/modals/user-create/user-create.component';
+
+/* ======================================== */
+/* INTERFACES */
+/* ======================================== */
+
+/**
+ * Interface para definir uma permissão de módulo
+ */
+interface Permission {
+  module: string;
+  access: boolean;
+  edit: boolean;
+}
+
+/**
+ * Interface para definir as opções da modal
+ */
+interface ModalOptions {
+  title?: string;
+  permissions: Permission[];
+  showUserFields?: boolean;
+}
+
+/* ======================================== */
+/* COMPONENTE: ACCESS AND SECURITY PAGE */
+/* ======================================== */
 
 @Component({
   selector: 'app-access-and-security',
@@ -12,7 +39,14 @@ import { IonContent, IonTitle, IonText, IonList, IonListHeader, IonItem, IonButt
 })
 export class AccessAndSecurityPage implements OnInit {
 
-  /**  */
+  /* ======================================== */
+  /* PROPRIEDADES */
+  /* ======================================== */
+
+  /** Controlador de modais do Ionic */
+  private modalCtrl: ModalController = inject(ModalController);
+
+  /** Lista de usuários da página */
   users = [
     {
       username: 'fulano.de.tal',
@@ -20,7 +54,6 @@ export class AccessAndSecurityPage implements OnInit {
       lastAccess: '14/07/2025 14:59',
       accessRemote: true,
       permissions: true,
-
     },
     {
       username: 'ciclana.pereira',
@@ -41,6 +74,76 @@ export class AccessAndSecurityPage implements OnInit {
   constructor() { }
 
   ngOnInit() {
+    // Inicialização da página
   }
 
+  /**
+ * Abre o modal de criação de usuário com opções personalizadas
+ * Segue o padrão estabelecido: array de opções + onDidDismiss()
+ */
+  async openUserCreateModal(): Promise<void> {
+    // Array de opções para a modal (padrão solicitado pelo gerente)
+    const modalOptions: ModalOptions = {
+      title: 'Adicionar Usuário',
+      permissions: [
+        { module: 'PROJETOS', access: true, edit: false },
+        { module: 'LOGÍSTICA', access: true, edit: false }
+      ],
+      showUserFields: true
+    };
+
+    // Criar e apresentar a modal com as opções
+    const modal = await this.modalCtrl.create({
+      component: UserCreateComponent,
+      componentProps: {
+        modalOptions: modalOptions
+      },
+      cssClass: 'user-create-modal'
+    });
+
+    // Apresentar a modal
+    await modal.present();
+
+    // Capturar o resultado usando onDidDismiss() (padrão solicitado)
+    const { data, role } = await modal.onDidDismiss();
+
+    // Processar o resultado da modal
+    if (data && data.success) {
+      console.log('Modal fechada com sucesso:', data);
+
+      // Verificar se há dados de usuário para adicionar à lista
+      if (data.data) {
+        console.log('Usuário criado:', data.data);
+        // Adicionar o novo usuário à lista
+        this.addNewUser(data.data);
+      }
+    } else {
+      console.log('Modal fechada sem ação');
+    }
+  }
+
+  /* ======================================== */
+  /* MÉTODOS PRIVADOS */
+  /* ======================================== */
+
+  /**
+   * Adiciona um novo usuário à lista de usuários
+   * @param userData - Dados do usuário criado
+   */
+  private addNewUser(userData: any): void {
+    // Criar novo usuário com dados da modal
+    const newUser = {
+      username: userData.username || 'novo.usuario',
+      fullName: userData.collaboratorName || 'Novo Usuário',
+      lastAccess: 'Agora',
+      accessRemote: true,
+      permissions: true,
+    };
+
+    // Adicionar à lista de usuários
+    this.users.unshift(newUser);
+
+    console.log('Novo usuário adicionado:', newUser);
+    console.log('Lista atualizada:', this.users);
+  }
 }
